fourth commit

+++ FibonacciApp/FibonacciSequence.swift
@@ -6,14 +6,12 @@
 //  Copyright (c) 2015 Shaun Chua. All rights reserved.
 //
 
-import Foundation
-
 class FibonacciSequence {
     
     let includesZero: Bool
-    let values: [UInt]
+    let values: [UInt64]
     
-    init(maxNumber: UInt, includesZero: Bool) {
+    init(maxNumber: UInt64, includesZero: Bool) {
         self.includesZero = includesZero
         
         if maxNumber == 0 && includesZero == false {
@@ -23,13 +21,13 @@ class FibonacciSequence {
             values = [0]
         }
         else {
-            var sequence:[UInt] = [0,1,1]
-            var nextNumber:UInt = 2
+            var sequence:[UInt64] = [0,1,1]
+            var nextNumber:UInt64 = 2
             while nextNumber <= maxNumber {
                 sequence.append(nextNumber)
                 let lastNumber = sequence.last!
                 let secondToLastNumber = sequence[sequence.count-2]
-                let (sum, didOverflow) = UInt.addWithOverflow(lastNumber, secondToLastNumber)
+                let (sum, didOverflow) = UInt64.addWithOverflow(lastNumber, secondToLastNumber)
                 if didOverflow == true {
                     println("Overflow! The next number is too large!")
                     break
@@ -43,7 +41,7 @@ class FibonacciSequence {
         }
     }
     
-    init(numberOfItemsInSequence: UInt, includesZero: Bool) {
+    init(numberOfItemsInSequence: UInt64, includesZero: Bool) {
         self.includesZero = includesZero
         
         if numberOfItemsInSequence == 0 {
@@ -58,7 +56,7 @@ class FibonacciSequence {
             }
         }
         else {
-            var sequence: [UInt]
+            var sequence: [UInt64]
             if includesZero == true {
                 sequence = [0,1]
             }
@@ -68,7 +66,7 @@ class FibonacciSequence {
             for var i = 2; i < Int(numberOfItemsInSequence); i++ {
                 let lastNumber = sequence.last!
                 let secondToLastNumber = sequence[sequence.count-2]
-                let (sum, didOverflow) = UInt.addWithOverflow(lastNumber, secondToLastNumber)
+                let (sum, didOverflow) = UInt64.addWithOverflow(lastNumber, secondToLastNumber)
                 if didOverflow == true {
                     println("Overflow! The next number is too large!")
                     break
